pyproject.toml
[project]
name = "get2-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.9",
    "aiosignal==1.3.1",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "asgiref==3.8.1",
    "attrs==24.2.0",
    "blinker==1.9.0",
    "branca==0.8.0",
    "cachetools==6.1.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "distlib==0.3.8",
    "distro==1.9.0",
    "dj-database-url>=3.0.1",
    "django==5.1.1",
    "django-cors-headers>=4.8.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "filelock==3.16.1",
    "folium==0.18.0",
    "frozenlist==1.4.1",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "gunicorn>=23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.27.2",
    "idna==3.10",
    "jinja2==3.1.4",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.5",
    "langgraph>=0.6.7",
    "markupsafe==3.0.2",
    "multidict==6.1.0",
    "narwhals==1.47.0",
    "numpy==2.1.3",
    "openai==1.52.2",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "platformdirs==4.3.6",
    "protobuf==6.31.1",
    "psycopg2-binary>=2.9.10",
    "pyarrow==20.0.0",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydeck==0.9.1",
    "python-dateutil==2.9.0.post0",
    "python-decouple>=3.8",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rest-framework-simplejwt>=0.0.2",
    "rpds-py==0.26.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlparse==0.5.1",
    "streamlit==1.46.1",
    "supabase==2.6.0", # 최신 버전으로 고정
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.66.5",
    "typing-extensions>=4.13.2", # 버전 업그레이드
    "tzdata==2024.1",
    "urllib3==2.2.3",
    "uv==0.7.20",
    "uvicorn[standard]==0.35.0",
    "virtualenv==20.26.5",
    "watchdog==6.0.0",
    "watchfiles==1.1.0",
    "websockets>=11,<15", # 버전 범위 수정
    "xyzservices==2024.9.0",
    "yarl==1.13.1",
    "chromadb>=0.5.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.1.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.8.0",
    "mcp>=1.0.0",
    "pydantic-settings>=2.5.0",
]

